envoy: Bazel rebase to 0.8.0, cache archive.
Bazel 0.8.0 changed the naming of directories in bazel-out symlink,
e.g., from "local-fastbuild" to "k8-fastbuild". "k8" here is the
default CPU target for Linux, which used to be "local"
before. Apparently this prefix will be different for different
targets, so the makefile should not assume to know them.

The reason for using those paths was that the "bazel-bin" symlink may
be changed when running bazel. If two targets are built (e.g.,
"release" and "debug") these will produce two different binaries, but
the "bazel-bin" will only lead to last of them (in the build order).

Bazels build directories are also dependent on the current user, so
they will be different for "vagrant" and "root" users. This is why
"install" target, that needs to be run as root, should not run bazel.

This patch solves this dilemma by simplifying the Makefile to only
ever produce a single envoy binary.

Makefile is also modified to explicitly check that the build targets
are not run as root, and that the install target is run as root.

Finally, support for archiving the bazel cache is added:

$ make bazel-archive - Build Envoy and tests using the current bazel
                       build cache, remove the Envoy and test binaries
                       and archive the bazel archive to
                       ~/bazel-cache.tar.bz2.

$ make bazel-clean-archive - Clear the current cache, build Envoy and
                             tests from scratch, remove the Envoy and
                             test binaries and archive the resulting
                             bazel archive to ~/bazel-cache.tar.bz2.

$ make bazel-restore - Clear the current cache and restore it from
                       ~/bazel-cache.tar.bz2.

After 'bazel-restore' the build (e.g., make tests) should only link
Envoy and test binaries. 'bazel-archive' can result in a bigger
archive, as it can carry on files that are strictly not needed any
more.

Envoronment variable BAZEL_ARCHIVE can be set to override the default
archive file path.

Signed-off-by: Jarno Rajahalme <jarno@covalent.io>