metrics: Correctly read endpoint_count metric
We previously incremented and decremented a gauge when we added and
removed endpoints. Unfortunately, the Insert/Delete functions for
endpoints were called gratuitiously and would result in double counting,
often pushing the count negative! This change collectes the count at the
time of a metrics read and should more correctly reflect the
point-in-time count.

Signed-off-by: Ray Bejjani <ray@covalent.io>